submarine/api/requirements.txt
# API Submarine ESP32 - Requirements
Flask==3.0.0
Flask-CORS==4.0.0
psutil==5.9.8
requests==2.31.0
pyserial==3.5
Werkzeug==3.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
